Add basic ee telemetry metrics
