+++ packages/graphql-server/src/ee/graphql/ee.graphql
@@ -1,5 +1,5 @@
 # import BuildImageJob, BuildImageOrderByInput, BuildImageJobConnection, BuildImageJobWhereInput, BuildImageJobWhereUniqueInput, BuildImage, BuildImageConnection, BuildImageWhereUniqueInput, BuildImageWhereInput, BuildImageCreateInput, BuildImageUpdateInput from "./buildImage.graphql"
-# import PhJob, PhJobOrderByInput, PhJobConnection, PhJobWhereUniqueInput, PhJobWhereInput, PhJobCreateInput, PhJobCancelResponse from "./phJob.graphql"
+# import PhJob, PhJobOrderByInput, PhJobConnection, PhJobWhereUniqueInput, PhJobWhereInput, PhJobCreateInput, PhJobCancelResponse, PhJobNotifyEventInput from "./phJob.graphql"
 # import PhSchedule, PhScheduleOrderByInput, PhScheduleConnection, PhScheduleWhereUniqueInput, PhScheduleWhereInput, PhScheduleCreateInput, PhScheduleUpdateInput, PhScheduleDeleteResponse, PhScheduleRunResponse from "./phSchedule.graphql"
 # import PhDeployment, PhDeploymentConnection, PhDeploymentWhereUniqueInput, PhDeploymentWhereInput, PhDeploymentCreateInput, PhDeploymentUpdateInput, PhDeploymentClientCreateInput, PhDeploymentClientCreateResponse, PhDeploymentClientWhereUniqueInput from "./phDeployment.graphql"
 # import UsageReport, UsageReportOrderByInput, UsageReportConnection, UsageReportWhereInput from "./usageReport.graphql"
@@ -138,6 +138,7 @@ type Mutation {
   rerunPhJob(where: PhJobWhereUniqueInput!): PhJob!
   cancelPhJob(where: PhJobWhereUniqueInput!): PhJobCancelResponse!
   cleanupPhJobArtifact: Int
+  notifyPhJobEvent(data: PhJobNotifyEventInput!): Int
 
   """PhSchedule"""
   createPhSchedule(data: PhScheduleCreateInput!): PhSchedule!
