+++ packages/graphql-server/src/utils/telemetryTraits.ts
@@ -2,7 +2,6 @@ import {TraitMiddleware} from './telemetry';
 import KcAdminClient from 'keycloak-admin';
 import { Config } from '../config';
 import { pickBy } from 'lodash';
-import CrdClient from '../crdClient/crdClientImpl';
 
 const userCount = async (kcAdminClient: KcAdminClient): Promise<number> => {
   return kcAdminClient.users.makeRequest({
@@ -21,28 +20,10 @@ const groupCount = async (kcAdminClient: KcAdminClient): Promise<number> => {
   });
 };
 
-const licenseStatusString = async (config: Config): Promise<string> => {
-  const status = config.licenseStatus.toLowerCase();
-  return status;
-};
-
-const modelDploymentCount = async (crdClient: CrdClient): Promise<number> => {
-  return (await crdClient.phDeployments.list()).length;
-};
-
-const scheduleCount = async (crdClient: CrdClient): Promise<number> => {
-  return (await crdClient.phSchedules.list()).length;
-};
-
 interface CreateDefaultTraitMiddlewareParams {
   config: Config;
   createKcAdminClient: () => KcAdminClient;
   getAccessToken: () => Promise<string>;
-}
-
-interface CreateEETraitMiddlewareParams {
-  config: Config;
-  crdClient: CrdClient;
 }
 
 export const createDefaultTraitMiddleware = (params: CreateDefaultTraitMiddlewareParams): TraitMiddleware => {
@@ -62,22 +43,6 @@ export const createDefaultTraitMiddleware = (params: CreateDefaultTraitMiddlewar
       ...dimensions,
       users,
       groups,
-    });
-    next();
-  };
-};
-
-export const createEETraitMiddleware = (params: CreateEETraitMiddlewareParams): TraitMiddleware => {
-  const {config, crdClient} = params;
-
-  return async (traits, next) => {
-    const licenseStatus = await licenseStatusString(config);
-    const deployments = await modelDploymentCount(crdClient);
-    const schedules = await scheduleCount(crdClient);
-    Object.assign(traits, {
-      licenseStatus,
-      deployments,
-      schedules,
     });
     next();
   };
